feat(example): log provider state
