+++ .changeset/provider-logging-example.md
@@ -2,4 +2,4 @@
 "example": patch
 ---
 
-Add provider logging showing window.ethereum state and EIP-6963 connectors. Sanitize logged connector data to avoid circular references.
+Add provider logging showing `window.ethereum` state and EIP-6963 connectors. Logs the Wagmi connectors array and sanitizes connector data to avoid circular references.
