+++ packages/example/src/pages/index.tsx
@@ -14,7 +14,6 @@ import React, { type ComponentProps, useEffect, useState } from 'react';
 import { type Address, parseEther } from 'viem';
 import {
   useAccount,
-  useConnectors,
   useSendTransaction,
   useSignMessage,
   useSignTypedData,
@@ -478,7 +477,6 @@ const Example = ({ authEnabled }: AppContextProps) => {
             </table>
           </div>
           {connected ? <ManageTransactions /> : null}
-          <ConnectorLogger />
         </>
       )}
     </div>
@@ -547,53 +545,6 @@ function ManageTransactions() {
         </button>
       </div>
     </form>
-  );
-}
-
-function ConnectorLogger() {
-  const connectors = useConnectors();
-  // Filter for EIP-6963 connectors (they will have rdns property)
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const eip6963Connectors = connectors.filter(
-    (connector) => (connector as any).rdns,
-  );
-
-  const connectorsInfo = eip6963Connectors.map((connector) => ({
-    id: connector.id,
-    name: connector.name,
-    ready: connector.ready,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    rdns: (connector as any).rdns,
-  }));
-
-  const ethereumInfo =
-    typeof window === 'undefined'
-      ? { defined: false, isMetaMask: false, isRainbow: false }
-      : {
-          defined: !!window.ethereum,
-          isMetaMask: (window as any).ethereum?.isMetaMask === true,
-          isRainbow: (window as any).ethereum?.isRainbow === true,
-        };
-
-  useEffect(() => {
-    console.log('window.ethereum:', ethereumInfo);
-  }, [ethereumInfo]);
-
-  useEffect(() => {
-    console.log('EIP-6963 connectors:', connectorsInfo);
-  }, [connectorsInfo]);
-
-  return (
-    <div style={{ fontFamily: 'monospace', marginTop: 16 }}>
-      <h3>Provider State</h3>
-      <pre>
-        {JSON.stringify(
-          { ethereum: ethereumInfo, connectors: connectorsInfo },
-          null,
-          2,
-        )}
-      </pre>
-    </div>
   );
 }
